reorganized project structure and fix tests
